Add files via upload
Market Basket analysis or affinity analysis, is a technique used by businesses to uncover patterns and relationships between items that are frequently purchased together by customers. It is based on the concept that certain products or items are often bought in combination, indicating a strong association or preference among customers.